Move CountryPill to it's own class
