+++ osu.Game.Tests/Visual/Online/TestSceneRankingsCountryFilter.cs
@@ -19,6 +19,7 @@ namespace osu.Game.Tests.Visual.Online
         public override IReadOnlyList<Type> RequiredTypes => new[]
         {
             typeof(CountryFilter),
+            typeof(CountryPill)
         };
 
         public TestSceneRankingsCountryFilter()
